Add compatibility with Python 2.4
2.4 has an incompatible syntax for binding exceptions to objects:

    try:
        something broken
    except: SyntaxError, e:
        handle(e)

2.6 and higher support this:

    try:
        something broken
    except: SynaxError as e:
        handle(e)

Python 3x dropped the former syntax so there isn't a unified way to bind the
caught exception via syntax.

Luckily, catching the exception without binding it to a variable is still the
same:

    try:
        something broken
    except: SyntaxError:
        pass

So we've added a custom function that looks up the message of the latest
exception and use that.

Also, Python 2.4 doesn't have a global `exit` function so we're using
`sys.exit` instead.